django_samples/requirements.txt
aioredis==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.11.0
django-crontab==0.7.1
django-redis==5.0.0
django-rest-framework-mongoengine==3.4.1
django-rq==2.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gevent==21.12.0
greenlet==1.1.2
h11==0.14.0
hiredis==2.0.0
httpcore==0.16.1
httpx==0.23.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
interchange==2021.0.4
mongoengine==0.23.1
monotonic==1.6
msgpack==0.6.2
numpy==1.23.1
optionaldict==0.1.2
packaging==21.3
pandas==1.4.3
pansi==2020.7.3
pdfminer.six==20221105
pdfplumber==0.7.6
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.13.0
PyJWT==2.4.0
pymongo==3.12.0
PyMuPDF==1.21.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rq==1.11.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sqlparse==0.4.2
tencentcloud-sdk-python==3.0.778
tencentcloud-sdk-python-ocr==3.0.778
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.20.0
Wand==0.6.10
websocket-client==1.2.3
wechatpy==1.8.18
wrapt==1.14.1
xmltodict==0.13.0
zope.event==4.5.0
zope.interface==5.4.0
django-filter==22.1
suds==1.1.2
celery==5.2.7

